+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.2.7] - Demo Apps - 2023-12-27
+### Added
+- Support of Alipay redirections and APK download in demo
+
+## [1.2.0] - Payments SDK - 2023-12-27
+### Added
+- Support of Alipay redirections and APK download in WebView
+
 ## [1.1.0] - Payments SDK - 2023-11-17
 ### Added
 - Custom Tabs support for the Huawei browser